fix: Address review feedback on emacs keybindings PR
- Fix Alt+F to skip whitespace before advancing through word (standard
  emacs forward-word behavior)
- Add detect_autocomplete_trigger() calls to Ctrl+K, Ctrl+U, Ctrl+W,
  and Ctrl+D to prevent stale autocomplete state
- Move emacs keybinding tests to separate file per CLAUDE.md convention
- Add test for Alt+F whitespace skip behavior
